reduce size of gui.rs
